feat: extract WalletSwitcherDropdown to dedicated component (temporary)
Moves inline WalletSwitcherDropdown composable from HomeScreen into
ui/components/WalletSwitcherDropdown.kt. HomeScreen now imports from
ui.components. Temporary until replaced by AccountSelectorSheet in Phase 2 Task 16.
